[Silver III] Title: 주유소, Time: 320 ms, Memory: 38796 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -23,14 +23,14 @@ public class Main {
             L_Arr[i] = Integer.parseInt(LInput[i]);
         }
 
-        int total = 0;
+        long total = 0;
         int nowIdx = 0;
         for (int i = 0; i < N-1; i++) {
             if (L_Arr[nowIdx] > L_Arr[i+1]) {
-                total += L_Arr[nowIdx] * kmArr[i];
+                total += (long) L_Arr[nowIdx] * kmArr[i];
                 nowIdx = i+1;
             } else {
-                total += L_Arr[nowIdx] * kmArr[i];
+                total += (long) L_Arr[nowIdx] * kmArr[i];
             }
         }
